feat : 기존의 main에 존재하던 isReady를 private로 만들기 위해 UserInterface 생성
